+++ core-repository/src/main/java/com/hmoa/core_repository/Perfume/PerfumeRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package com.hmoa.core_repository
+package com.hmoa.core_repository.Perfume
 
 import com.hmoa.core_datastore.Perfume.PerfumeDataStore
 import com.hmoa.core_model.request.AgeRequestDto
